backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-cors-headers==4.3.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
filetype==1.2.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.1.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.4.4
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pep8-naming==0.13.3
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.25.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
PyMuPDF==1.23.16
PyMuPDFb==1.23.9
pyparsing==3.1.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
rsa==4.9
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
views-py==2.0.0
